commands: move files code to files.rs
